Move tests in checkatlas package
