--- Containerfile
@@ -1,5 +1,0 @@
-FROM python:3.7-alpine
-COPY . /app
-WORKDIR /app
-RUN pip install .
-CMD ["checkatlas"]
