+++ Makefile
@@ -33,12 +33,11 @@ lint:             ## Run pep8, black, mypy linters.
 	$(ENV_PREFIX)poetry run flake8 checkatlas/
 	$(ENV_PREFIX)poetry run flake8 tests/
 	$(ENV_PREFIX)poetry run black -l 79 --check checkatlas/
-	$(ENV_PREFIX)poetry run black -l 79 --check tests/
 	$(ENV_PREFIX)poetry run mypy --ignore-missing-imports checkatlas/
 
 .PHONY: test
 test:             ## Run tests and generate coverage report.
-	$(ENV_PREFIX)poetry run pytest -v --cov-config .coveragerc --cov=checkatlas -l --tb=short --maxfail=1 tests/
+	$(ENV_PREFIX)poetry run pytest -v --cov-config .coveragerc --cov=checkatlas -l --tb=short --maxfail=1 checkatlas/tests/
 	$(ENV_PREFIX)poetry run coverage xml
 	$(ENV_PREFIX)poetry run coverage html
 
@@ -59,7 +58,7 @@ ci:          ## Run a continuous integration : Add every change to git and creat
 	@git add --all
 	@git commit -m "Continuous integration 🔄 tests-$(VERSION)"
 	@echo "creating git tag : tests-$(VERSION)"
-	@git tag tests-$(VERSION)-11
+	@git tag tests-$(VERSION)-12
 	@git push -u origin HEAD --tags
 	@echo "Github Actions will detect the new tag and run the continuous integration process."
