+++ checkatlas/tests/test_atlas.py
@@ -2,27 +2,40 @@ import argparse
 import os
 
 import pytest
-from rpy2.robjects.methods import RS4
+from anndata import AnnData
 
-from checkatlas import atlas, atlas_seurat, checkatlas
+from checkatlas import atlas, checkatlas
 from checkatlas.utils import files, folders
 
-from tests import datasets
+from checkatlas.tests import datasets
 
 given = pytest.mark.parametrize
 
 
-@given("atlas_path,expected", [(datasets.SEURAT_TEST_PATH, RS4)])
-def test_read_seurat_atlas(atlas_path, expected):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
-    print(type(seurat_data))
-    assert type(seurat_data) == expected
+@given("atlas_path,expected", [(datasets.ADATA_TEST_PATH, AnnData)])
+def test_read_scanpy_atlas(atlas_path, expected):
+    adata = atlas.read_atlas(atlas_path)
+    assert type(adata) == expected
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path,expected", [(datasets.CELLRANGER_TEST_PATH, AnnData)])
+def test_read_cellranger_atlas(atlas_path, expected):
+    adata = atlas.read_cellranger(atlas_path)
+    assert type(adata) == expected
+
+
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
+def test_clean_scanpy(atlas_path):
+    adata = atlas.read_atlas(atlas_path)
+    adata_clean = atlas.clean_scanpy_atlas(adata, atlas_path)
+    list_var = adata_clean.var_names
+    assert len(set(list_var)) == len(list_var)
+
+
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_viable_obs_qc(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
-    expected = ["nCount_RNA", "nFeature_RNA"]
+    adata = atlas.read_atlas(atlas_path)
+    expected = ["n_genes_by_counts", "total_counts"]
     parser = argparse.ArgumentParser()
     parser.add_argument("--qc_display")
     args = parser.parse_args(
@@ -36,42 +49,41 @@ def test_viable_obs_qc(atlas_path):
             ],
         ]
     )
-    obs_keys = atlas_seurat.get_viable_obs_qc(seurat_data, args)
+    obs_keys = atlas.get_viable_obs_qc(adata, args)
     assert obs_keys == expected
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_viable_obs_annot(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
-    expected = ["RNA_snn_res.0.5", "seurat_clusters"]
+    adata = atlas.read_atlas(atlas_path)
+    expected = ["leiden", "louvain"]
     parser = argparse.ArgumentParser()
     parser.add_argument("--obs_cluster")
     args = parser.parse_args(["--obs_cluster", atlas.OBS_CLUSTERS])
-    obs_keys = atlas_seurat.get_viable_obs_annot(seurat_data, args)
+    obs_keys = atlas.get_viable_obs_annot(adata, args)
     assert obs_keys == expected
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_viable_obsm(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
-    expected = ["pca", "umap", "tsne"]
+    adata = atlas.read_atlas(atlas_path)
+    expected = ["X_draw_graph_fr", "X_pca", "X_tsne", "X_umap"]
     parser = argparse.ArgumentParser()
     parser.add_argument("--obs_cluster")
     args = parser.parse_args(["--obs_cluster", atlas.OBS_CLUSTERS])
-    obs_keys = atlas_seurat.get_viable_obsm(seurat_data, args)
-    print(obs_keys)
+    obs_keys = atlas.get_viable_obsm(adata, args)
     assert obs_keys == expected
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_summary_table(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     checkatlas_path = os.getcwd()
     args = parser.parse_args(["--path", checkatlas_path])
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_summary_table(seurat_data, atlas_path, args)
+    atlas.create_summary_table(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.SUMMARY, checkatlas.SUMMARY_EXTENSION, args.path
@@ -79,15 +91,15 @@ def test_summary_table(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
-def test_seurat_data_table(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
+def test_adata_table(atlas_path):
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     checkatlas_path = os.getcwd()
     args = parser.parse_args(["--path", checkatlas_path])
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_anndata_table(seurat_data, atlas_path, args)
+    atlas.create_anndata_table(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.ANNDATA, checkatlas.ADATA_EXTENSION, args.path
@@ -96,9 +108,9 @@ def test_seurat_data_table(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_qc_table(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--qc_display")
@@ -120,7 +132,7 @@ def test_qc_table(atlas_path):
         ]
     )
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_qc_tables(seurat_data, atlas_path, args)
+    atlas.create_qc_tables(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.QC, checkatlas.QC_EXTENSION, args.path
@@ -128,15 +140,15 @@ def test_qc_table(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_qc_plots(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     checkatlas_path = os.getcwd()
     args = parser.parse_args(["--path", checkatlas_path])
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_qc_plots(seurat_data, atlas_path, args)
+    atlas.create_qc_plots(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.QC_FIG, checkatlas.QC_FIG_EXTENSION, args.path
@@ -144,9 +156,9 @@ def test_qc_plots(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_umap_plots(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--obs_cluster")
@@ -156,7 +168,7 @@ def test_umap_plots(atlas_path):
     )
 
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_umap_fig(seurat_data, atlas_path, args)
+    atlas.create_umap_fig(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.UMAP, checkatlas.UMAP_EXTENSION, args.path
@@ -164,9 +176,9 @@ def test_umap_plots(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_tsne_plots(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--obs_cluster")
@@ -176,7 +188,7 @@ def test_tsne_plots(atlas_path):
     )
 
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.create_tsne_fig(seurat_data, atlas_path, args)
+    atlas.create_tsne_fig(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name, folders.TSNE, checkatlas.TSNE_EXTENSION, args.path
@@ -184,9 +196,9 @@ def test_tsne_plots(atlas_path):
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_cluster_metric(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--obs_cluster")
@@ -203,7 +215,7 @@ def test_cluster_metric(atlas_path):
         ]
     )
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.metric_cluster(seurat_data, atlas_path, args)
+    atlas.metric_cluster(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
     csv_path = files.get_file_path(
         atlas_name,
@@ -214,43 +226,61 @@ def test_cluster_metric(atlas_path):
     assert os.path.exists(csv_path)
 
 
-""" @given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_annot_metric(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--obs_cluster")
     parser.add_argument("--metric_annot")
     checkatlas_path = os.getcwd()
-    args = parser.parse_args(['--path',checkatlas_path,
-                              '--metric_annot',
-                                ["rand_index"],
-                                "--obs_cluster",atlas.OBS_CLUSTERS])
+    args = parser.parse_args(
+        [
+            "--path",
+            checkatlas_path,
+            "--metric_annot",
+            ["rand_index"],
+            "--obs_cluster",
+            atlas.OBS_CLUSTERS,
+        ]
+    )
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.metric_annot(seurat_data, atlas_path, args)
+    atlas.metric_annot(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
-    csv_path = files.get_file_path(atlas_name, folders.ANNOTATION,
-                                   checkatlas.METRIC_ANNOTATION_EXTENSION,
-                                   args.path)
+    csv_path = files.get_file_path(
+        atlas_name,
+        folders.ANNOTATION,
+        checkatlas.METRIC_ANNOTATION_EXTENSION,
+        args.path,
+    )
     assert os.path.exists(csv_path)
 
 
-@given("atlas_path", [(datasets.SEURAT_TEST_PATH)])
+@given("atlas_path", [(datasets.ADATA_TEST_PATH)])
 def test_dimred_metric(atlas_path):
-    seurat_data = atlas_seurat.read_atlas(atlas_path)
+    adata = atlas.read_atlas(atlas_path)
     parser = argparse.ArgumentParser()
     parser.add_argument("--path")
     parser.add_argument("--obs_cluster")
     parser.add_argument("--metric_dimred")
     checkatlas_path = os.getcwd()
-    args = parser.parse_args(['--path',checkatlas_path,
-                              '--metric_dimred',
-                                ["kruskal_stress"],
-                                "--obs_cluster",atlas.OBS_CLUSTERS])
+    args = parser.parse_args(
+        [
+            "--path",
+            checkatlas_path,
+            "--metric_dimred",
+            ["kruskal_stress"],
+            "--obs_cluster",
+            atlas.OBS_CLUSTERS,
+        ]
+    )
     folders.checkatlas_folders(checkatlas_path)
-    atlas_seurat.metric_dimred(seurat_data, atlas_path, args)
+    atlas.metric_dimred(adata, atlas_path, args)
     atlas_name = checkatlas.get_atlas_name(atlas_path)
-    csv_path = files.get_file_path(atlas_name, folders.DIMRED,
-                                   checkatlas.METRIC_DIMRED_EXTENSION,
-                                   args.path)
-    assert os.path.exists(csv_path) """
+    csv_path = files.get_file_path(
+        atlas_name,
+        folders.DIMRED,
+        checkatlas.METRIC_DIMRED_EXTENSION,
+        args.path,
+    )
+    assert os.path.exists(csv_path)
