+++ pyproject.toml
@@ -7,7 +7,7 @@ readme = "README.md"
 license = "BSD 3-Clause"
 packages = [{include = "checkatlas"}]
 include = ["checkatlas/checkatlas_workflow.nf"]
-exclude = ["tests/data/", ".github", "*dask-worker-space*"]
+exclude = ["checkatlas/tests/data/", ".github", "*dask-worker-space*"]
 homepage = "https://checkatlas.readthedocs.io/"
 repository = "https://github.com/becavin-lab/checkatlas/"
 documentation = "https://checkatlas.readthedocs.io/"
